requirements.txt
flask==3.0.3
python-dotenv>=1.0,<2.0
playwright>=1.50,<2.0
requests>=2.31,<3.0
